requirements.txt
mcp>=1.0.0
openai>=1.60.0
msgraph-sdk>=1.0.0
azure-identity>=1.16.0
python-dotenv>=1.0.0
httpx>=0.27.0
